refactor and added function to show graph
